Restore COPYING to GPL2, and stop updating it
The rng-tools code is licensed according to the GPL v2 or any
later version of the GPL.

Due to a major thinko, autogen.sh was forcing the COPYING file to be
updated, and it ended up containg the GPL v3 text in version
2-unofficial-mt.12, instead of the correct GPL v2 text.

All source code files correctly state that they are distributed under
the GPL v2 or any later version.

Testore the COPYING file to contain the text of the GPL v2, and change
autogen.sh to not update it anymore.