viapadlock: work around xstore CPU errata
It has been reported that on at least one Nano CPU the xstore instruction
will write as many as 16 bytes of data to the output buffer.  This is an
extremely dangerous CPU bug.

Give xstore an overkill 128-byte buffer aligned to a 16-byte boundary.
That ought to keep the VIA Nano happy.

Closes: Debian #609289