rdrand: Build rngd with libgcrypt where available
Partial port of upstream commit 04361df2dc2add0ab085ad77eb9681fd42bd9cca.

The data from the x86 RDRAND instruction needs to be conditioned before
it can be used as a seed source.  This is normally done using the AES-NI
instructions, but on systems that have RDRAND but not AES-NI, libgcrypt
is used instead.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>